Adding functional test projects for provider-agnostic tests Using project references instead of source linking Making the optimistic concurrency tests provider-agnostic

+++ src/EntityFramework.SQLite/SQLiteDataStoreCreator.cs
@@ -111,7 +111,7 @@ namespace Microsoft.Data.Entity.SQLite
 
         public override async Task<bool> HasTablesAsync(CancellationToken cancellationToken = new CancellationToken())
         {
-            return (int)(await _executor.ExecuteScalarAsync(_connection.DbConnection, CreateHasTablesCommand(), cancellationToken)) != 0;
+            return (long)(await _executor.ExecuteScalarAsync(_connection.DbConnection, CreateHasTablesCommand(), cancellationToken)) != 0;
         }
 
         private SqlStatement CreateHasTablesCommand()
